+++ mediation/Chartboost/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Chartboost Unity Mediation Plugin Changelog
 
+#### Version Next
+- Renamed `DummyClient` to `PlaceholderClient`.
+
 #### [Version 4.6.0](https://dl.google.com/googleadmobadssdk/mediation/unity/chartboost/ChartboostUnityAdapter-4.6.0.zip)
 - Supports [Chartboost Android adapter version 9.7.0.0](https://github.com/googleads/googleads-mobile-android-mediation/blob/main/ThirdPartyAdapters/chartboost/CHANGELOG.md#version-9700).
 - Supports [Chartboost iOS adapter version 9.7.0.0](https://github.com/googleads/googleads-mobile-ios-mediation/blob/main/adapters/Chartboost/CHANGELOG.md#version-9700).